pyproject.toml
[project]
name = "bookstoreflaskapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]
